Move natural breaks function to data file so other data sets don't have to import census response
